Feature/save interfaces (#1)
* Provide Stream Saver Interfaces

* Add IIOStreamProvider file implementation

* Add IAsyncIOStreamProvider http implementation

Currently missing some constructor parameters for HttpClient initialization.

* Add IStreamSerializer and IAsyncStreamSerializer implementation for Json

* Rewrite SaveFile to use interfaces.

Additionally, expose a static class SaveFile with the method CreateGZipJsonFile. This returns a SaveFile using File Input/Output, Json Serialization and GZip compression. This is likely how the SaveFile will mostly be used.

* parameter should be interface instead of class

* Add documentation

* Exhaustive overloads for SaveFile.CreateGZipJsonFile

Now you can call this method with a JsonSerializerContext or JsonTypeInfo parameter.

* Functionally tested HttpIO

The HttpIO is ready for the field! It now exposes the necessary headers and optional Request uris, some easy-to-use constructors and properly disposes its objects.

* Remove null-forgiving operators from deserialize extension methods

* Remove ConfigureAwait(false) from all async calls as it introduces save issues

* Fix ContentLength = 0 on Write

* Add namespaces for Compression, IO, and Serialization

* Fix IAsyncIOStreamProvider

We need to compress the memory stream and leave it open BEFORE writing it with our IAsyncIOStreamProvider.

Note that this isn't a completely correct name, since it doesn't provide a write / input stream, it accepts one. Will need to change this once a good name has been devised.

* Add HttpIO documentation

* Fix naming in static SaveFile

Gzip should be GZip