Rewrite Godot tests as unit tests
The dependency on Godot for running tests is unnecessary, as all of the functionality in this library is based on native C#. All of the tests have been rewritten to use xunit for quicker testing. Moq has been used for testing the SaveFile class and its many possible configurations based on the given interfaces.